# openapi_generated.yaml
openapi: 3.0.3
info:
  title: Agri Clinic API
  version: 1.0.0
  description: "Production-ready REST API for the Agri Clinic SaaS platform.\n\n##
    Authentication\nAll protected endpoints require a **Bearer JWT** token.  \nObtain
    tokens via `POST /api/v1/auth/login/` then pass:  \n`Authorization: Bearer <access_token>`\n\n##
    Standard Response Envelope\n**Success:** `{ success: true, message: string, data:
    object }`  \n**Error:**   `{ success: false, message: string, errors: object,
    code: string }`\n\n## Pagination\nAll list endpoints return `{ count, next, previous,
    results }` under `data`.\nUse `?page=N&page_size=N` query params."
paths:
...
  /api/v1/admin/visits/:
    get:
      operationId: v1_admin_visits_list
      description: Admin can only view visits. Visits are created by employees.
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminVisitList'
          description: ''
  /api/v1/admin/visits/{id}/:
    get:
      operationId: v1_admin_visits_retrieve
      description: Admin can only view visits. Visits are created by employees.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVisit'
          description: ''
...
  /api/v1/farmers/{farmer_id}/visits/:
    get:
      operationId: v1_farmers_visits_list
      description: Returns paginated visit history for the specified farmer.
      summary: List visits for a farmer
      parameters:
      - in: path
        name: farmer_id
        schema:
          type: integer
        required: true
      - in: query
        name: page
        schema:
          type: integer
        description: Page number (1-based).
      - in: query
        name: page_size
        schema:
          type: integer
        description: Number of results per page (default 50, max varies).
      tags:
      - Visits
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FarmerVisit'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmerVisitNotFound'
          description: ''
...
  /api/v1/mobile/visits/:
    get:
      operationId: v1_mobile_visits_list
      description: GET returns all visits for the logged-in employee. POST creates
        a visit and optionally uploads media files.
      summary: Mobile visits list/create
      tags:
      - Mobile
      - Visits
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileVisitValidationError'
          description: ''
    post:
      operationId: v1_mobile_visits_create
      description: GET returns all visits for the logged-in employee. POST creates
        a visit and optionally uploads media files.
      summary: Mobile visits list/create
      tags:
      - Mobile
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileVisitValidationError'
          description: ''
...
  /api/v1/visits/:
    get:
      operationId: v1_visits_list
      description: Paginated list of all visits ordered by visit date, newest first.
        Employees see all visits.
      summary: List visits
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: Page number (1-based).
      - in: query
        name: page_size
        schema:
          type: integer
        description: Number of results per page (default 50, max varies).
      tags:
      - Visits
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
          description: ''
    post:
      operationId: v1_visits_create
      description: Create a new field visit. Media files can be attached as multipart
        `media_files[]`.
      summary: Create visit
      tags:
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitCreateError'
          description: ''
...
  /api/v1/visits/{id}/:
    get:
      operationId: v1_visits_retrieve
      description: Admin can only view visits. Visits are created by employees.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVisit'
          description: ''
    put:
      operationId: v1_visits_update
      description: Retrieve or update a visit by ID. PUT/PATCH supports full or partial
        updates.
      summary: Get or update visit details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              type: object
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
    patch:
      operationId: v1_visits_partial_update
      description: Retrieve or update a visit by ID. PUT/PATCH supports full or partial
        updates.
      summary: Get or update visit details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              type: object
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
...
  /api/v1/visits/visits/{id}:
    get:
      operationId: v1_visits_visits_retrieve
      description: Retrieve or update a visit by ID. PUT/PATCH supports full or partial
        updates.
      summary: Get or update visit details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Visits
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
    put:
      operationId: v1_visits_visits_update
      description: Retrieve or update a visit by ID. PUT/PATCH supports full or partial
        updates.
      summary: Get or update visit details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              type: object
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
    patch:
      operationId: v1_visits_visits_partial_update
      description: Retrieve or update a visit by ID. PUT/PATCH supports full or partial
        updates.
      summary: Get or update visit details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Visits
      requestBody:
        content:
          application/json:
            schema:
              type: object
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
...
components:
  schemas:
...
    AdminVisit:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        employee_name:
          type: string
          readOnly: true
          default: ''
        village_name:
          type: string
          readOnly: true
          default: ''
        crop_name:
          type: string
          readOnly: true
          default: ''
        issues:
          type: array
          items:
            $ref: '#/components/schemas/AdminCropIssue'
          readOnly: true
        media_files:
          type: array
          items:
            $ref: '#/components/schemas/AdminVisitMedia'
          readOnly: true
        visit_date:
          type: string
          format: date
          nullable: true
        visit_time:
          type: string
          format: time
          readOnly: true
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        address:
          type: string
          nullable: true
        farmer_name:
          type: string
          nullable: true
          maxLength: 255
        farmer_phone:
          type: string
          nullable: true
          maxLength: 20
        land_name:
          type: string
          nullable: true
          maxLength: 255
        land_area:
          type: number
          format: double
          nullable: true
        crop_stage:
          type: string
          nullable: true
          maxLength: 100
        variety:
          type: string
          nullable: true
          maxLength: 100
        season:
          type: string
          nullable: true
          maxLength: 100
        sowing_date:
          type: string
          format: date
          nullable: true
        crop_health:
          type: string
          nullable: true
          maxLength: 100
        pest_issue:
          type: boolean
          nullable: true
        disease_issue:
          type: boolean
          nullable: true
        weed_condition:
          type: string
          nullable: true
          maxLength: 100
        notes:
          type: string
          nullable: true
        fertilizer_advice:
          type: string
          nullable: true
        pesticide_advice:
          type: string
          nullable: true
        irrigation_advice:
          type: string
          nullable: true
        general_advice:
          type: string
          nullable: true
        follow_up_required:
          type: boolean
          nullable: true
        next_visit_date:
          type: string
          format: date
          nullable: true
        status:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/VisitStatusEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        employee:
          type: integer
        district:
          type: integer
          nullable: true
        village:
          type: integer
          nullable: true
        crop:
          type: integer
          nullable: true
      required:
      - created_at
      - crop_name
      - employee
      - employee_name
      - id
      - issues
      - media_files
      - updated_at
      - village_name
      - visit_time
...
    FarmerVisitNotFound:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        errors:
          type: object
          additionalProperties: {}
          default: {}
        code:
          type: string
      required:
      - code
      - message
...
    MobileVisitValidationError:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        errors:
          type: object
          additionalProperties: {}
          default: {}
        code:
          type: string
      required:
      - code
      - message
...
    PaginatedAdminVisitList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminVisit'
...
    VisitStatusEnum:
      enum:
      - active
      - completed
      - pending
      type: string
      description: |-
        * `active` - Active
        * `completed` - Completed
        * `pending` - Pending
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...